added active, completed and all filters
When in active mode, only active todos are shown, and possibility to add new. The color of rows is adjusted.
When in completed mode, only completed are presentd. Not possible to add new. Color of rows is adjusted.
When in all mode, all todos are presented. Possible to add new todos.